# document/api/api.yml
openapi: 3.0.3
info:
  title: PetStore API
  version: "0.1.1"
  description: |-
    ## PetStore OpenAPI 設計
    - バックエンド： Laravel
    - フロントエンド： Vue
  license:
    name: MIT
servers:
  - url: "http://localhost/api/"
    description: 開発サーバーURL
  - url: "https://product/api/"
    description: 本番サーバーURL
paths:
...
  /pet:
    post:
      operationId: addNewPet
      tags:
        - Pet
      summary: ペット新規登録
      description: |
        ペット新規登録
        - category は　categoriesから選択したID
        - tags : tagsテーブルにある場合、ID登録、ない場合 tagsテーブルに新規登録
        - イメージアップロードは　api/pet/uploadImage
      responses:
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/unexpected"
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/pet"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      requestBody:
        content:
          applicaiton/json:
            schema:
              $ref: "#/components/schemas/request_pet_store"
      security:
        - apiKey: []
        - bearer: []
  /pets:
    get:
      operationId: getAllPets
      summary: ペット一覧取得
      tags:
        - Pet
      description: |-
        stagusがavailableはorder可能
      responses:
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/unexpected"
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/response_pegination"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      parameters:
        - name: order
          schema:
            type: string
          in: query
        - name: sorted
          schema:
            type: string
          in: query
        - name: page
          schema:
            type: integer
          in: query
      security:
        - apiKey: []
  /pet/{petId}:
    get:
      operationId: getPetById
      tags:
        - Pet
      summary: ペット情報取得
      responses:
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                type: object
                $ref: "#/components/schemas/pet"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      parameters:
        - name: petId
          schema:
            type: integer
          in: path
          required: true
      security:
        - apiKey: []
    put:
      operationId: updatePetById
      tags:
        - Pet
      summary: ペット情報更新
      description: |-
        更新処理
      responses:
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/response_ok"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      parameters:
        - name: petId
          schema:
            type: integer
          in: path
          required: true
      requestBody:
        content:
          applicaiton/json:
            schema:
              $ref: "#/components/schemas/request_pet_update"
      security:
        - apiKey: []
        - bearer: []
    delete:
      operationId: deletePetById
      tags:
        - Pet
      summary: ペット情報削除
      responses:
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/response_ok"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      parameters:
        - name: petId
          schema:
            type: integer
          in: path
          required: true
      security:
        - apiKey: []
        - bearer: []
...
  /pet/uploadImage:
    post:
      tags:
        - Pet
      operationId: uploadImage
      summary: アップロードペットイメージ
      description: |-
        イメージファイルアップロード
        - tmpフォルダへ保存
        - pet 新規登録、修正する成功時、 petsフォルダへ移動
      responses:
        "200":
          description: OK
          content:
            applicaiton/json:
              schema:
                type: object
                properties:
                  file_name:
                    type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      security:
        - apiKey: []
        - bearer: []
...
components:
  schemas:
    pet_comment:
      type: object
      properties:
        id:
          type: integer
        pet_id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        comment:
          type: string
        created_at:
          type: string
    category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    pet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        photo_urls:
          type: array
          items:
            type: string
        category:
          type: object
          $ref: "#/components/schemas/category"
        owner:
          type: object
          $ref: "#/components/schemas/user"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/pet_comment"
    user:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
...
    request_pet_store:
      type: object
      required:
        - name
        - category
        - phpto_urls
      properties:
        name:
          type: string
        category_id:
          type: integer
        photo_urls:
          type: array
          maxItems: 20
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
    request_pet_update:
      type: object
      required:
        - id
        - name
        - category
      properties:
        id:
          type: integer
        name:
          type: string
        category_id:
          type: integer
        photo_urls:
          type: array
          maxItems: 20
          items:
            type: string
        tags:
          type: array
          maxItems: 20
          items:
            type: string
...
    response_ok:
      required:
        - ok
      type: object
      properties:
        ok:
          type: boolean
    response_pegination:
      type: object
      properties:
        current_page:
          type: integer
        first_page_url:
          type: string
        from:
          type: integer
        last_page:
          type: integer
        last_page_url:
          type: integer
        next_page_url:
          type: string
        path:
          type: string
        per_page:
          type: integer
        prev_page_url:
          type: string
        to:
          type: integer
        total:
          type: integer
        links:
          type: array
          maxItems: 10
          items:
            type: object
            properties:
              url:
                type: string
              label:
                type: string
              active:
                type: boolean
        data:
          type: array
          maxItems: 15
          items: {}
    error400:
      type: object
      properties:
        status:
          type: string
          default: error
        message:
          type: string
        code:
          type: integer
    error500:
      type: object
      properties:
        status:
          type: string
          default: error
        message:
          type: string
        code:
          type: integer
    unexpected:
      type: object
      properties:
        status:
          type: string
          default: error
        message:
          type: string
        code:
          type: integer
  securitySchemes:
    apiKey:
      type: apiKey
      name: API-KEY
      in: header
    bearer:
      type: http
      scheme: bearer
      bearerFormat: bearer
security:
  - apiKey: []
  - bearer: []